requirements.txt
flask
openai
pandas
python-dotenv
/sqlite3
requests